+++ src/main/java/org/tron/core/services/ratelimiter/strategy/QpsStrategy.java
@@ -12,19 +12,19 @@ public class QpsStrategy extends Strategy {
   private RateLimiter rateLimiter;
 
   public static final String STRATEGY_PARAM_QPS = "qps";
-  public static final int DEFAULT_QPS = 100;
+  public static final Double DEFAULT_QPS = 100D;
 
   // define the default strategy params
   @Override
   protected Map<String, ParamItem> defaultParam() {
     Map<String, ParamItem> map = new HashMap<>();
-    map.put(STRATEGY_PARAM_QPS, new ParamItem(Integer.class, DEFAULT_QPS));
+    map.put(STRATEGY_PARAM_QPS, new ParamItem(Double.class, DEFAULT_QPS));
     return map;
   }
 
   public QpsStrategy(String paramString) {
     super(paramString);
-    rateLimiter = RateLimiter.create((Integer) mapParams.get(STRATEGY_PARAM_QPS).value);
+    rateLimiter = RateLimiter.create((Double) mapParams.get(STRATEGY_PARAM_QPS).value);
   }
 
   public boolean acquire() {
